Preparing foldering struct for EC2 deployment
