+++ client/README.md
@@ -1,0 +1,15 @@
+# Getting Started with SIMS Dashboard
+
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+
+## Available Scripts
+
+- In the project directory, run `npm install` to install dependencies first.
+- Then you are ready to `npm start` to start development.
+
+### Preview of Build
+
+Production branch is set to `master` as default.
+Please checkout to `dev` branch first and pull then make your own branch from there if you are to contribute.
+
+Build preview is [here](https://new-sims-template-4cpaktiqh-aoponcedeleon.vercel.app/).
