+++ client/src/components/BaitStationMonitoring.tsx
