+++ client/src/components/DashboardShell.tsx
