+++ client/src/components/Findings.tsx
