+++ client/src/components/RealtimeMonitoring.tsx
