+++ client/src/modules/DeviceInspectionTable.tsx
