+++ package.json
@@ -1,63 +1,22 @@
 {
   "name": "new-sims-template",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@mantine/core": "^4.0.8",
-    "@mantine/dates": "^4.0.8",
-    "@mantine/hooks": "^4.0.8",
-    "@mantine/notifications": "^4.2.5",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.26",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
-    "axios": "^0.27.2",
-    "chart.js": "^3.7.1",
-    "dayjs": "^1.11.0",
-    "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
-    "react-chartjs-2": "^4.0.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "tabler-icons-react": "^1.41.0",
-    "typescript": "^4.6.3",
-    "web-vitals": "^2.1.4"
-  },
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/aoponcedeleon/new-sims-template.git"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/aoponcedeleon/new-sims-template/issues"
   },
-  "devDependencies": {
-    "@faker-js/faker": "^6.0.0",
-    "@typescript-eslint/parser": "^5.16.0",
-    "eslint": "^8.12.0",
-    "prettier": "^2.6.1",
-    "prettier-eslint": "^13.0.0"
+  "homepage": "https://github.com/aoponcedeleon/new-sims-template#readme",
+  "dependencies": {
+    "express": "^4.18.1"
   }
 }
